Calculate Suitability Indices

SIcalc computes suitability indices given a set of suitability curves and project-specific inputs. Suitability indices may be computed based on either linear interpolation (for continuous variables) or a lookup method (for categorical variables).

Usage

SIcalc(SI, input_proj)

Arguments

SI

data frame or matrix; matrix of suitability curves ordered as parameter breakpoints and associated suitability indices for each parameter.

input_proj

list; A list of numeric or categorical vectors of application-specific input parameters associated with the suitability curve data from SI.

Value

A vector of the suitability index values that match given user inputs. Values are returned as equal to the extreme of a range if inputs are outside of model range.

Details

The SI parameter must contain a set of column pairs. The column pairs should consist of a

References

US Fish and Wildlife Service. (1980). Habitat as a basis for environmental assessment. Ecological Services Manual, 101.

US Fish and Wildlife Service. (1980). Habitat Evaluation Procedures (HEP). Ecological Services Manual, 102.

US Fish and Wildlife Service. (1981). Standards for the Development of Habitat Suitability Index Models. Ecological Services Manual, 103.

Examples

# Build manual model
# Number of trees > 51cm diameter per 0.4 ha plot
tree.num     <- c(0, 2, 4, NA)              # parameter breakpoints
tree.num.SIV <- c(0.1, 1, 1, NA)            # parameter suitability indices
tree_num <- data.frame(tree.num, tree.num.SIV)

# Mean diameter of overstory trees
avg.dbh      <- c(0, 5, 20, NA)             # parameter breakpoints
avg.dbh.SIV  <- c(0, 0, 1, NA)              # parameter suitability indices
tree_diameter <- data.frame(avg.dbh, avg.dbh.SIV)

# Percent canopy cover of overstory trees
can.cov      <- c(0, 20, 60, 100)           # parameter breakpoints
can.cov.SIV  <- c(0, 0, 1, 1)               # parameter suitability indices
canopy_cov <- data.frame(can.cov, can.cov.SIV)

barredowl <- data.frame(tree_num, tree_diameter, canopy_cov)

bo_test1 <- list(2, 5, 20)

# Barredowl si values
bo_si_1 <- nybem::SIcalc(barredowl, bo_test1)